+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - markdown output to playground
 - document catalog filter by input type on playground
+- system info in playground
 
 ### Changed